add BuildDB model
This is used to abstract the Django ORM from the rest of the
application.

+++ src/gentoo_build_publisher/views.py
@@ -1,16 +1,14 @@
 """
 View for gbp
 """
-from django.db.models import Count
 from django.http import HttpRequest, HttpResponse, JsonResponse
-from django.shortcuts import get_object_or_404
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 
 from gentoo_build_publisher.build import Build
+from gentoo_build_publisher.db import BuildDB
 from gentoo_build_publisher.diff import dirdiff
 from gentoo_build_publisher.managers import BuildMan
-from gentoo_build_publisher.models import BuildLog, BuildModel
 from gentoo_build_publisher.tasks import publish_build, pull_build
 
 
@@ -41,37 +39,42 @@ def show_build(
     _request: HttpRequest, build_name: str, build_number: int
 ) -> JsonResponse:
     """View details of a build"""
-    build_model = get_object_or_404(BuildModel, name=build_name, number=build_number)
-    response = BuildMan(build_model).as_dict()
+    build = Build(name=build_name, number=build_number)
+    buildman = BuildMan(build)
+    response = buildman.as_dict()
     response["error"] = None
 
-    return JsonResponse(response)
+    status_code = 200 if buildman.db else 404
+
+    return JsonResponse(response, status=status_code)
 
 
 @require_POST
 @csrf_exempt
 def delete(_request: HttpRequest, build_name: str, build_number: int) -> JsonResponse:
     """View to delete a build"""
-    build_model = get_object_or_404(BuildModel, name=build_name, number=build_number)
-    buildman = BuildMan(build_model)
+    build = Build(name=build_name, number=build_number)
+    buildman = BuildMan(build)
 
-    buildman.delete()
+    if buildman.db:
+        buildman.delete()
+        return JsonResponse({"deleted": True, "error": None})
 
-    return JsonResponse({"deleted": True, "error": None})
+    return JsonResponse({"error": "Not found"}, status=404)
 
 
 def latest(_request: HttpRequest, build_name: str) -> JsonResponse:
     """View to return the latest build for a machine"""
-    builds = BuildModel.objects.filter(
-        name=build_name, completed__isnull=False
-    ).order_by("-submitted")
+    builds = BuildDB.builds(name=build_name, completed__isnull=False)
 
-    if builds.count() == 0:
+    try:
+        build_db = next(builds)
+    except StopIteration:
         return JsonResponse(
             {"error": "No completed builds exist with that name"}, status=404
         )
 
-    response = BuildMan(builds[0]).as_dict()
+    response = BuildMan(build_db).as_dict()
     response["error"] = None
 
     return JsonResponse(response)
@@ -79,11 +82,10 @@ def latest(_request: HttpRequest, build_name: str) -> JsonResponse:
 
 def list_builds(_request: HttpRequest, build_name: str) -> JsonResponse:
     """View to return the list of builds with the given machine"""
-    builds = BuildModel.objects.filter(
-        name=build_name, completed__isnull=False
-    ).order_by("number")
+    builds = list(BuildDB.builds(name=build_name, completed__isnull=False))
+    builds.sort(key=lambda i: i.number)
 
-    if builds.count() == 0:
+    if not builds:
         return JsonResponse(
             {"error": "No completed builds exist with that name", "builds": []},
             status=404,
@@ -99,19 +101,28 @@ def list_builds(_request: HttpRequest, build_name: str) -> JsonResponse:
 
 def logs(_request: HttpRequest, build_name: str, build_number: int) -> HttpResponse:
     """View to return the Jenkins build logs for a given build"""
-    build_log = get_object_or_404(
-        BuildLog, build_model__name=build_name, build_model__number=build_number
-    )
+    build = Build(name=build_name, number=build_number)
+    build_db = BuildDB.get(build)
 
-    return HttpResponse(build_log.logs, content_type="text/plain")
+    if build_db is None or build_db.logs is None:
+        return HttpResponse("Not Found", content_type="text/plain", status=404)
+
+    return HttpResponse(build_db.logs, content_type="text/plain")
 
 
 def diff_builds(
     _request: HttpRequest, build_name: str, left: int, right: int
 ) -> JsonResponse:
     """View to show the diff between two builds for the given machines"""
-    left_build = BuildMan(get_object_or_404(BuildModel, name=build_name, number=left))
-    right_build = BuildMan(get_object_or_404(BuildModel, name=build_name, number=right))
+    left_build = BuildMan(Build(name=build_name, number=left))
+
+    if not left_build.db:
+        return JsonResponse({"error": "left build not found"}, status=404)
+
+    right_build = BuildMan(Build(name=build_name, number=right))
+
+    if not right_build.db:
+        return JsonResponse({"error": "right build not found"}, status=404)
 
     left_path = left_build.storage_build.get_path(Build.Content.BINPKGS)
     right_path = right_build.storage_build.get_path(Build.Content.BINPKGS)
@@ -131,10 +142,7 @@ def diff_builds(
 
 def list_machines(_request: HttpRequest) -> JsonResponse:
     """List the machines and build counts"""
-    machines = [
-        *BuildModel.objects.values("name")
-        .order_by("name")
-        .annotate(builds=Count("name"))
-    ]
+    machines = BuildDB.list_machines()
+    machines.sort(key=lambda i: i.name)
 
-    return JsonResponse({"error": None, "machines": machines})
+    return JsonResponse({"error": None, "machines": [i.as_dict() for i in machines]})
